FinalProject/services/face_recognition_service/requirements-aaca.txt
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dlib==20.0.0
dotenv==0.9.9
ecdsa==0.19.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
flask-cors==6.0.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
pgvector==0.4.1
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.1.0
python-jose==3.3.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
